+++ CHANGES.md
@@ -70,6 +70,7 @@
 * Cythonized radial sensitivity to shear/bulk functions in `TidalPy.tides.multilayer.sensitivity` (based on Tobie+2005)
 * Cythonized radial heating functions that use the sensitivity to shear/bulk functions in `TidalPy.tides.multilayer.heating` (based on Tobie+2005)
 * Improved logging so it is less spamy.
+* Logger now logs all exceptions raised.
 
 #### Dependencies
 * Added support for CyRK v0.12.x